+++ lib/test_screen/component_test_screen.dart
@@ -9,8 +9,8 @@ import 'package:recipe_app/presentation/component/input_field.dart';
 import 'package:recipe_app/presentation/component/tabs.dart';
 import 'package:recipe_app/presentation/saved_recipes/saved_recipes_screen.dart';
 import 'package:recipe_app/presentation/saved_recipes/saved_recipes_view_model.dart';
-import 'package:recipe_app/presentation/search_recipes.dart/search_recipes_screen.dart';
-import 'package:recipe_app/presentation/search_recipes.dart/search_recipes_view_model.dart';
+import 'package:recipe_app/presentation/search_recipes/search_recipes_screen.dart';
+import 'package:recipe_app/presentation/search_recipes/search_recipes_view_model.dart';
 import 'package:recipe_app/presentation/splash/splash_screen.dart';
 import 'package:recipe_app/test_screen/button_screen.dart';
 import 'package:recipe_app/test_screen/ingredient_screen.dart';
